+++ lib/ui/WebLoginNetIdPanel.dart
@@ -24,6 +24,8 @@ import 'package:illinois/service/Analytics.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:illinois/ui/onboarding2/Onboarding2Widgets.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
+import 'package:rokwire_plugin/utils/utils.dart';
+import 'package:web/web.dart' as web;
 
 class WebLoginNetIdPanel extends StatefulWidget {
   WebLoginNetIdPanel();
@@ -121,7 +123,8 @@ class _WebLoginNetIdPanelState extends State<WebLoginNetIdPanel> {
     Analytics().logSelect(target: 'Log in with NetID');
     if (_loginProgress != true) {
       setState(() { _loginProgress = true; });
-      Auth2().authenticateWithOidc().then((Auth2OidcAuthenticateResult? result) {
+      web.Window? webWindow = WebUtils.createIosWebWindow();
+      Auth2().authenticateWithOidc(iosWebWindow: webWindow).then((Auth2OidcAuthenticateResult? result) {
         if (mounted) {
           if (result == Auth2OidcAuthenticateResult.succeeded) {
             FlexUI().update().then((_) {