+++ lib/ui/illini/WordlePanel.dart
@@ -4,17 +4,29 @@ import 'dart:math';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/service/Content.dart';
+import 'package:http/http.dart';
+import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/service/Config.dart';
 import 'package:illinois/service/Storage.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
+import 'package:illinois/utils/AppUtils.dart';
 import 'package:intl/intl.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:rokwire_plugin/ext/network.dart';
 import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
+import 'package:uuid/uuid.dart';
 
 class WordlePanel extends StatefulWidget {
+  final WordleGame? game;
+  final WordleDailyWord? dailyWord;
+  final Set<String>? dictionary;
+
+  WordlePanel({super.key, this.game, this.dailyWord, this.dictionary});
+
   @override
   State<StatefulWidget> createState() => _WordlePanelState();
 }
@@ -32,7 +44,9 @@ class _WordlePanelState extends State<WordlePanel> with NotificationsListener {
     NotificationService().subscribe(this, [
       WordleGameWidget.notifyGameOver,
     ]);
+
     _initData();
+
     super.initState();
   }
 
@@ -73,7 +87,7 @@ class _WordlePanelState extends State<WordlePanel> with NotificationsListener {
             Text(Localization().getStringEx('panel.wordle.heading.info.text', 'Presented by The Daily Illini'), style: Styles().textStyles.getTextStyleEx('widget.message.light.small'), textAlign: TextAlign.center,)
           ),
           Padding(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 16), child:
-            AspectRatio(aspectRatio: 1, child:
+            AspectRatio(aspectRatio: _dailyWord?.asectRatio ?? 1.0, child:
               _bodyContent
             ),
           ),
@@ -110,18 +124,41 @@ class _WordlePanelState extends State<WordlePanel> with NotificationsListener {
   // Data
 
   Future<void> _initData() async {
-    setState(() {
-      _loadProgress = true;
-    });
 
-    List<dynamic> results = await Future.wait([
-      WordleGame.loadDailyWord(),
-      WordleGame.loadDictionary(),
-    ]);
+    WordleDailyWord? dailyWord = widget.dailyWord;
+    Set<String>? dictionary = widget.dictionary;
 
-    WordleGame? game = WordleGame.fromStorage();
-    WordleDailyWord? dailyWord = JsonUtils.cast(ListUtils.entry(results, 0));
-    Set<String>? dictionary = JsonUtils.setStringsValue(ListUtils.entry(results, 1));
+    List<Future<dynamic>> futures = <Future<dynamic>>[];
+
+    int dailyWordIndex = (dailyWord == null) ? futures.length : -1;
+    if (0 <= dailyWordIndex) {
+      futures.add(WordleGame.loadDailyWord());
+    }
+
+    int dictionaryIndex = (dictionary == null) ? futures.length : -1;
+    if (0 <= dictionaryIndex) {
+      futures.add(WordleGame.loadDictionary());
+    }
+
+
+    if (0 < futures.length) {
+
+      setState(() {
+        _loadProgress = true;
+      });
+
+      List<dynamic> results = await Future.wait(futures);
+
+      if (0 <= dailyWordIndex) {
+        dailyWord = JsonUtils.cast(ListUtils.entry(results, dailyWordIndex));
+      }
+
+      if (0 <= dictionaryIndex) {
+        dictionary = JsonUtils.setStringsValue(ListUtils.entry(results, dictionaryIndex));
+      }
+    }
+
+    WordleGame? game = widget.game ??  WordleGame.fromStorage();
     if ((dailyWord != null) && ((game == null) || (game.word != dailyWord.word))) {
       game = WordleGame(dailyWord.word);
     }
@@ -133,6 +170,19 @@ class _WordlePanelState extends State<WordlePanel> with NotificationsListener {
         _dictionary = dictionary;
         _loadProgress = false;
       });
+
+      _logFinishedAlert(game);
+    }
+
+  }
+
+  // Analytics
+
+  void _logFinishedAlert(WordleGame? game) {
+    if ((game != null) && game.isFinished) {
+      Analytics().logAlert(text: game.isSucceeded ?
+        Localization().getStringEx('widget.wordle.game.status.succeeded.title', 'You win!', language: 'en') :
+        Localization().getStringEx('widget.wordle.game.status.failed.title', 'You lost', language: 'en'));
     }
   }
 
@@ -140,11 +190,13 @@ class _WordlePanelState extends State<WordlePanel> with NotificationsListener {
 
   void _onGameOver(WordleGame? game) {
     if ((game != null) && mounted) {
+      _logFinishedAlert(game);
       setState(() {
         _game = game;
       });
     }
   }
+
 
   void _onLongPressLogo() {
     if (_dailyWord != null) {
@@ -164,13 +216,14 @@ class _WordlePanelState extends State<WordlePanel> with NotificationsListener {
   }
 }
 
-class WordleWidget extends StatelessWidget {
+class WordleWidget extends StatefulWidget {
 
   final WordleGame game;
   final WordleDailyWord dailyWord;
   final Set<String>? dictionary;
   final bool autofocus;
   final bool hintMode;
+  final double gutterRatio;
 
   WordleWidget({super.key,
     required this.game,
@@ -178,34 +231,66 @@ class WordleWidget extends StatelessWidget {
     this.dictionary,
     this.autofocus = false,
     this.hintMode = false,
+    this.gutterRatio = 0.075,
   });
 
   @override
-  Widget build(BuildContext context) => game.isFinished ?
+  State<StatefulWidget> createState() => _WordleWidgetState();
+}
+
+class _WordleWidgetState extends State<WordleWidget> {
+
+  bool _gameStatusContentEnabled = true;
+
+  @override
+  void didUpdateWidget(covariant WordleWidget oldWidget) {
+    if ((widget.game != oldWidget.game) && mounted) {
+      _gameStatusContentEnabled = true;
+    }
+    super.didUpdateWidget(oldWidget);
+  }
+
+  @override
+  Widget build(BuildContext context) => widget.game.isFinished ?
     _gameStatusContent : _wordleGameContent;
 
-  Widget get _wordleGameContent => WordleGameWidget(game, dictionary: dictionary, autofocus: autofocus, hintMode: hintMode);
-  Widget get _wordlePreviewContent => WordleGameWidget(game, enabled: false,);
+  Widget get _wordleGameContent => WordleGameWidget(widget.game, dictionary: widget.dictionary, autofocus: widget.autofocus, hintMode: widget.hintMode, gutterRatio: widget.gutterRatio);
+  Widget get _wordlePreviewContent => WordleGameWidget(widget.game, enabled: false, gutterRatio: widget.gutterRatio);
   Widget get _wordlePreviewLayer =>  Container(color: Styles().colors.blackTransparent018,);
 
   Widget get _gameStatusContent =>
     Stack(children: [
       _wordlePreviewContent,
-      Positioned.fill(child: _wordlePreviewLayer),
-      Positioned.fill(child:
-        Align(alignment: Alignment.bottomCenter, child:
-          Padding(padding: EdgeInsets.symmetric(horizontal: 8, vertical: 8), child:
-            _gameStatusPopup,
-          )
-        )
-      ),
+      if (_gameStatusContentEnabled)
+        ...[
+          Positioned.fill(child: _wordlePreviewLayer),
+          Positioned.fill(child:
+            Align(alignment: Alignment.bottomCenter, child:
+              Padding(padding: EdgeInsets.symmetric(horizontal: 8, vertical: 8), child:
+                _gameStatusPopup,
+              )
+            )
+          ),
+        ]
     ],);
 
 
-  Widget get _gameStatusPopup {
-    return Container(decoration: _gameStatusPopupDecoration, padding: _gameStatusPopupPadding, child:
+  Widget get _gameStatusPopup =>
+    Container(decoration: _gameStatusPopupDecoration, child:
+      Stack(children: [
+        _gameStatusPopupContent,
+        Positioned.fill(child:
+          Align(alignment: Alignment.topRight, child:
+            _gameStatusPopupCloseButton
+          )
+        ),
+      ],)
+    );
+
+  Widget get _gameStatusPopupContent =>
+    Padding(padding: _gameStatusPopupPadding, child:
       Column(mainAxisSize: MainAxisSize.min, children: [
-        Text(game.isSucceeded ?
+        Text(widget.game.isSucceeded ?
             Localization().getStringEx('widget.wordle.game.status.succeeded.title', 'You win!') :
             Localization().getStringEx('widget.wordle.game.status.failed.title', 'You lost'),
           style: _gameStatusTitleTextStyle, textAlign: TextAlign.center,
@@ -213,17 +298,18 @@ class WordleWidget extends StatelessWidget {
 
         Padding(padding: EdgeInsets.only(top: 12), child:
           Column(mainAxisSize: MainAxisSize.min, children: [
-            Text(Localization().getStringEx('widget.wordle.game.status.word.text', 'Today\'s word: {{word}}').replaceAll('{{word}}', dailyWord.word.toUpperCase()),
+            Text(Localization().getStringEx('widget.wordle.game.status.word.text', 'Today\'s word: {{word}}').replaceAll('{{word}}', widget.dailyWord.word.toUpperCase()),
               style: _gameStatusSectionTextStyle, textAlign: TextAlign.center,
             ),
-            Text(DateFormat(Localization().getStringEx('widget.wordle.game.status.date.text.format', 'MMMM, dd, yyyy')).format(dailyWord.dateTime ?? DateTime.now()),
-              style: _gameStatusInfoTextStyle, textAlign: TextAlign.center,
-            ),
-            if (game.isSucceeded && (dailyWord.author?.isNotEmpty == true))
-              Text(Localization().getStringEx('widget.wordle.game.status.author.text', 'Edited by {{author}}').replaceAll('{{author}}', dailyWord.author ?? ''),
+            if (widget.dailyWord.dateUtc != null)
+              Text(DateFormat(Localization().getStringEx('widget.wordle.game.status.date.text.format', 'MMMM, dd, yyyy')).format(widget.dailyWord.dateUtc?.toLocal() ?? DateTime.now()),
                 style: _gameStatusInfoTextStyle, textAlign: TextAlign.center,
               ),
-            if (game.isSucceeded && (dailyWord.quote?.isNotEmpty == true))
+            if (widget.game.isSucceeded && (widget.dailyWord.author?.isNotEmpty == true))
+              Text(Localization().getStringEx('widget.wordle.game.status.author.text', 'Edited by {{author}}').replaceAll('{{author}}', widget.dailyWord.author ?? ''),
+                style: _gameStatusInfoTextStyle, textAlign: TextAlign.center,
+              ),
+            if (widget.game.isSucceeded && (widget.dailyWord.storyTitle?.isNotEmpty == true))
               ...[
                 Padding(padding: EdgeInsets.only(top: 12), child:
                   Text(Localization().getStringEx('widget.wordle.game.status.related_to.text', 'Related to this word'),
@@ -231,18 +317,32 @@ class WordleWidget extends StatelessWidget {
                   ),
                 ),
                 Padding(padding: EdgeInsets.only(top: 2, bottom: 4), child:
-                  Container(decoration: _gameStatusQuoteDecoration, padding: _gameStatusQuotePadding, child:
-                    Text(dailyWord.quote ?? '',
-                      style: _gameStatusInfoTextStyle, textAlign: TextAlign.center,
+                  _gameStatusStoryWrapper(child:
+                    Container(decoration: _gameStatusStoryDecoration, padding: _gameStatusStoryPadding, child:
+                      Text(widget.dailyWord.storyTitle ?? '',
+                        style: _gameStatusInfoTextStyle, textAlign: TextAlign.center,
+                      ),
                     ),
-                  )
+                  ),
                 )
               ]
           ]),
         ),
       ],)
     );
-  }
+
+  Widget _gameStatusStoryWrapper({ required Widget child }) =>
+    (widget.dailyWord.storyUrl?.isNotEmpty == true) ?
+      InkWell(onTap: () => _onTapStatusStory(context), child: child,) : child;
+
+  Widget get _gameStatusPopupCloseButton =>
+    Semantics(label: Localization().getStringEx('dialog.close.title', 'Close'), button: true, excludeSemantics: true, child:
+      InkWell(onTap : _onStatusPopupClose, child:
+        Padding(padding: EdgeInsets.all(14), child:
+          Styles().images.getImage('close-circle-small', excludeFromSemantics: true)
+        ),
+      ),
+    );
 
   Decoration get _gameStatusPopupDecoration => BoxDecoration(
     color: Styles().colors.surface,
@@ -256,28 +356,43 @@ class WordleWidget extends StatelessWidget {
   TextStyle? get _gameStatusSectionTextStyle => Styles().textStyles.getTextStyle('widget.message.regular.fat');
   TextStyle? get _gameStatusInfoTextStyle => Styles().textStyles.getTextStyle('widget.message.regular');
 
-  Decoration get _gameStatusQuoteDecoration => BoxDecoration(
+  Decoration get _gameStatusStoryDecoration => BoxDecoration(
     color: Styles().colors.lightGray,
     border: Border.all(color: Styles().colors.surfaceAccent2),
   );
-  EdgeInsetsGeometry get _gameStatusQuotePadding => EdgeInsets.symmetric(horizontal: 8, vertical: 4);
+  EdgeInsetsGeometry get _gameStatusStoryPadding => EdgeInsets.symmetric(horizontal: 8, vertical: 4);
+
+  void _onTapStatusStory(BuildContext context) {
+    Analytics().logSelect(target: 'Story Url');
+    AppLaunchUrl.launch(context: context, url: widget.dailyWord.storyUrl, tryInternal: false);
+  }
+
+  void _onStatusPopupClose() {
+    Analytics().logSelect(target: 'Close');
+    setState(() {
+      _gameStatusContentEnabled = false;
+    });
+  }
 }
 
 class WordleGameWidget extends StatefulWidget {
 
   static const String notifyGameOver = 'edu.illinois.rokwire.illini.wordle.game.over';
+  static const String notifyGameProgress = 'edu.illinois.rokwire.illini.wordle.game.progress';
 
   final WordleGame game;
   final Set<String>? dictionary;
   final bool enabled;
   final bool autofocus;
   final bool hintMode;
+  final double gutterRatio;
 
   WordleGameWidget(this.game, { super.key,
     this.dictionary,
     this.enabled = true,
     this.autofocus = false,
     this.hintMode = false,
+    this.gutterRatio = 0.075,
   });
 
   @override
@@ -311,7 +426,7 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
 
   @override
   void didUpdateWidget(covariant WordleGameWidget oldWidget) {
-    if (!identical(widget.game, oldWidget.game) && mounted) {
+    if ((widget.game != oldWidget.game) && mounted) {
       _moves = List<String>.from(widget.game.moves);
       _rack = '';
     }
@@ -330,7 +445,7 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
 
   Widget get _wordleWidget =>
     GestureDetector(behavior: HitTestBehavior.opaque, onTap: _onTapWordle, onLongPress: _onLongPressWordle, child:
-      AspectRatio(aspectRatio: 1, child:
+      AspectRatio(aspectRatio: widget.game.asectRatio, child:
         _buildWords()
       ),
     );
@@ -340,12 +455,17 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
     int wordIndex = 0;
     List<Widget> words = <Widget>[];
 
-    int movesCount = min(_moves.length, WordleGame.numberOfWords);
+    int movesCount = min(_moves.length, widget.game.numberOfWords);
     while (wordIndex < movesCount) {
       if (words.isNotEmpty) {
         words.add(Expanded(flex: _gutterFlex, child: Container()));
       }
-      words.add(Expanded(flex: _cellFlex, child: _buildWord(_moves[wordIndex], _WordleLetterDisplay.move)));
+      String move = _moves[wordIndex];
+      words.add(Expanded(flex: _cellFlex, child: _buildWord(
+        word: move,
+        status: widget.game.wordStatus(move),
+        display: _WordleLetterDisplay.move
+      )));
       wordIndex++;
     }
 
@@ -353,11 +473,15 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
       if (words.isNotEmpty) {
         words.add(Expanded(flex: _gutterFlex, child: Container()));
       }
-      words.add(Expanded(flex: _cellFlex, child: _buildWord(_rack, _WordleLetterDisplay.rack)));
+      words.add(Expanded(flex: _cellFlex, child: _buildWord(
+        word: _rack,
+        status: widget.game.wordStatus(_rack),
+        display: _WordleLetterDisplay.rack,
+      )));
       wordIndex++;
     }
 
-    while (wordIndex < WordleGame.numberOfWords) {
+    while (wordIndex < widget.game.numberOfWords) {
       if (words.isNotEmpty) {
         words.add(Expanded(flex: _gutterFlex, child: Container()));
       }
@@ -368,18 +492,22 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
     return Column(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: words);
   }
 
-  Widget _buildWord([String word = '', _WordleLetterDisplay display = _WordleLetterDisplay.rack]) {
+  Widget _buildWord({
+    String word = '',
+    List<_WordleLetterStatus> status = const <_WordleLetterStatus>[],
+    _WordleLetterDisplay display = _WordleLetterDisplay.rack
+  }) {
 
     int letterIndex = 0;
     List<Widget> letters = <Widget>[];
 
-    int wordLength = min(word.length, WordleGame.wordLength);
+    int wordLength = min(word.length, widget.game.wordLength);
     while (letterIndex < wordLength) {
       if (letters.isNotEmpty) {
         letters.add(Expanded(flex: _gutterFlex, child: Container()));
       }
       String letter = word.substring(letterIndex, letterIndex + 1);
-      _WordleLetterStatus? letterStatus = widget.game.letterStatus(letter, letterIndex);
+      _WordleLetterStatus? letterStatus = ListUtils.entry(status, letterIndex);
       letters.add(Expanded(flex: _cellFlex, child: _buildCell(
         letter: letter,
         status: letterStatus,
@@ -389,7 +517,7 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
       letterIndex++;
     }
 
-    while (letterIndex < WordleGame.wordLength) {
+    while (letterIndex < widget.game.wordLength) {
       if (letters.isNotEmpty) {
         letters.add(Expanded(flex: _gutterFlex, child: Container()));
       }
@@ -400,7 +528,12 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
     return Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: letters);
   }
 
-  Widget _buildCell({String letter = '', _WordleLetterStatus? status, _WordleLetterDisplay display = _WordleLetterDisplay.rack, bool hintMode = false }) {
+  Widget _buildCell({
+    String letter = '',
+    _WordleLetterStatus? status,
+    _WordleLetterDisplay display = _WordleLetterDisplay.rack,
+    bool hintMode = false
+  }) {
 
     Color backColor = ((display == _WordleLetterDisplay.move) && (status != null)) ? status.color : Styles().colors.surface;
     TextStyle? textStyle = ((display == _WordleLetterDisplay.move) && (status != null)) ?
@@ -415,22 +548,18 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
       border: Border.all(color: borderColor, width: borderWidth)
     );
 
-
-    return AspectRatio(aspectRatio: 1, child:
-      Container(decoration: cellDecoration, child:
-        Padding(padding: EdgeInsets.all(8), child:
-          Center(child:
-            Text(letter.toUpperCase(), style: textStyle,)
-          )
-        ),
+    return Container(decoration: cellDecoration, child:
+      Padding(padding: EdgeInsets.all(8), child:
+        Center(child:
+          Text(letter.toUpperCase(), style: textStyle,)
+        )
       ),
     );
   }
 
-  static const double _gutter = 0.075;
   static const double _gutterPrec = 1000;
-  int get _gutterFlex => (_gutter * _gutterPrec).toInt();
-  int get _cellFlex => ((1 - _gutter) * _gutterPrec).toInt();
+  int get _gutterFlex => (widget.gutterRatio * _gutterPrec).toInt();
+  int get _cellFlex => ((1 - widget.gutterRatio) * _gutterPrec).toInt();
 
 
   // Keyboard
@@ -489,20 +618,21 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
   void _onLongPressWordle() {
 
   }
+
   // Rack
 
   void _onKeyCharacter(String character) {
-    debugPrint('Key: $character');
-    if (character.isAlpha && (_rack.length < WordleGame.wordLength) && mounted) {
+    //debugPrint('Key: $character');
+    if (character.isAlpha && (_rack.length < widget.game.wordLength) && mounted) {
       setState(() {
-        _rack = _rack + character.toLowerCase();
+        _rack = _rack + character.toUpperCase();
       });
     }
     _textFocusNode.requestFocus(); // show again
   }
 
   void _onBackward() {
-    debugPrint('Backward');
+    //debugPrint('Backward');
     if (_rack.isNotEmpty && mounted) {
       setState(() {
         _rack = _rack.substring(0, _rack.length - 1);
@@ -512,25 +642,33 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
   }
 
   void _onSubmitWord() {
-    debugPrint('Submit');
-    if (_rack.length == WordleGame.wordLength) {
-      if ((widget.dictionary?.isNotEmpty == true) && (widget.dictionary?.contains(_rack) != true)) {
+    //debugPrint('Submit');
+    if (_rack.length == widget.game.wordLength) {
+      if ((widget.dictionary?.isNotEmpty == true) && (Storage().debugWordleIgnoreDictionary != true) && (widget.dictionary?.contains(_rack) != true)) {
+        _logAnalytics(_rack, _moves.length + 1, status: AnalyticsIllordleEventStatus.notInDictionary);
         AppToast.showMessage(Localization().getStringEx('widget.wordle.move.invalid.text', 'Not in word list'), gravity: ToastGravity.CENTER, duration: Duration(milliseconds: 1000));
         _textFocusNode.requestFocus(); // show again
       }
-      else if (_moves.length < WordleGame.numberOfWords) {
+      else if (_moves.length < widget.game.numberOfWords) {
         setState(() {
           _moves.add(_rack);
           _rack = '';
         });
 
         WordleGame game = WordleGame.fromOther(widget.game, moves: _moves);
-        Storage().illordleGame = game.toStorageString();
+        game.saveToStorage();
 
-        if ((_moves.last == widget.game.word) || (_moves.length == WordleGame.numberOfWords)) {
+        if (_moves.last == widget.game.word) {
+          _logAnalytics(_moves.last, _moves.length, status: AnalyticsIllordleEventStatus.success);
+          NotificationService().notify(WordleGameWidget.notifyGameOver, game);
+        }
+        if (_moves.length == widget.game.numberOfWords) {
+          _logAnalytics(_moves.last, _moves.length, status: AnalyticsIllordleEventStatus.fail);
           NotificationService().notify(WordleGameWidget.notifyGameOver, game);
         }
         else {
+          _logAnalytics(_moves.last, _moves.length);
+          NotificationService().notify(WordleGameWidget.notifyGameProgress, game);
           _textFocusNode.requestFocus(); // show again
         }
       }
@@ -543,39 +681,73 @@ class _WordleGameWidgetState extends State<WordleGameWidget> {
     }
   }
 
+  void _logAnalytics(String guess, int attempt, { AnalyticsIllordleEventStatus? status }) =>
+      Analytics().logIllordle(
+        word: widget.game.word,
+        guess: guess,
+        attempt: attempt,
+        status: status,
+      );
+
 }
 
 class WordleGame {
-  static const int wordLength = 5;
-  static const int numberOfWords = 5;
 
+  final String uuid;
   final String word;
-  final Set<String> _wordChars;
   final List<String> moves;
 
-  WordleGame(this.word, { this.moves = const <String>[]}) :
-    _wordChars = Set<String>.from(word.characters);
+  WordleGame(this.word, { String? uuid, this.moves = const <String>[] }) :
+    this.uuid = uuid ?? Uuid().v1();
 
   factory WordleGame.fromOther(WordleGame other, {
     String? word,
+    String? uuid,
     List<String>? moves,
   }) => WordleGame(word ?? other.word,
+    uuid: uuid ?? other.uuid,
     moves: moves ?? other.moves,
   );
 
   // Accessories
 
+  int get wordLength => word.wordLength;
+  int get numberOfWords => word.numberOfWords;
+  double get asectRatio => word.asectRatio;
+
   bool get isSucceeded => moves.isNotEmpty && (moves.last == word);
   bool get isFailed => (moves.length == numberOfWords) && (moves.last != word);
   bool get isFinished => isSucceeded || isFailed;
 
-  _WordleLetterStatus letterStatus(String letter, [int position = 0]) {
-    if (_wordChars.contains(letter)) {
-      return ((0 <= position) && (position < word.length) && (word.substring(position, position + 1) == letter)) ? _WordleLetterStatus.inPlace : _WordleLetterStatus.inUse;
+  List<_WordleLetterStatus> wordStatus(String guess) {
+
+    Map<String, int> wordLettersBag = _LettersBag.fromWord(word);
+    List<_WordleLetterStatus> guessStatus = List<_WordleLetterStatus>.filled(guess.length, _WordleLetterStatus.outOfUse);
+
+    // Pass 1: determine inPlace
+    for (int index = 0; index < guess.length; index++) {
+      if (index < word.length) {
+        String guessLetter = guess.substring(index, index + 1);
+        String wordLetter = word.substring(index, index + 1);
+        if (guessLetter == wordLetter) {
+          guessStatus[index] = _WordleLetterStatus.inPlace;
+          wordLettersBag.removeLetter(wordLetter);
+        }
+      }
     }
-    else {
-      return _WordleLetterStatus.outOfUse;
+
+    // Pass 2: determine inUse
+    for (int guessIndex = 0; guessIndex < guess.length; guessIndex++) {
+      if (guessStatus[guessIndex] == _WordleLetterStatus.outOfUse) {
+        String guessLetter = guess.substring(guessIndex, guessIndex + 1);
+        if (wordLettersBag.containsLetter(guessLetter)) {
+          guessStatus[guessIndex] = _WordleLetterStatus.inUse;
+          wordLettersBag.removeLetter(guessLetter);
+        }
+      }
     }
+
+    return guessStatus;
   }
 
   // Equality
@@ -584,65 +756,105 @@ class WordleGame {
   bool operator==(Object other) =>
     (other is WordleGame) &&
     (word == other.word) &&
+    (uuid == other.uuid) &&
     DeepCollectionEquality().equals(moves, other.moves);
 
   @override
   int get hashCode =>
     word.hashCode ^
+    uuid.hashCode ^
     DeepCollectionEquality().hash(moves);
+
+  // Json Serialization
+
+  static WordleGame? fromJson(Map<String, dynamic>? json) {
+    String? word = JsonUtils.stringValue(json?['word']);
+    String? uuid = JsonUtils.stringValue(json?['uuid']);
+    List<String>? moves = JsonUtils.listStringsValue(json?['moves']);
+    return ((word != null) && (uuid != null) && (moves != null)) ? WordleGame(word, uuid: uuid, moves: moves) : null;
+  }
+
+  Map<String, dynamic> toJson() => <String, dynamic>{
+    'word': word,
+    'uuid': uuid,
+    'moves': moves,
+  };
 
   // Storage Serialization
 
   static WordleGame? fromStorage() =>
-      WordleGame.fromStorageString(Storage().illordleGame);
-
-  static WordleGame? fromStorageString(String? value) {
-    List<String> entries = (value != null) ? value.split(_storageDelimiter) : <String>[];
-    return (1 < entries.length) ? WordleGame(entries.first, moves: entries.sublist(1)) : null;
-  }
+    WordleGame.fromJson(JsonUtils.decodeMap(Storage().wordleGame) );
 
   void saveToStorage() =>
-    Storage().illordleGame = toStorageString();
-
-  String toStorageString() => <String>[
-    word,
-    ...moves
-  ].join(_storageDelimiter);
-
-  static const String _storageDelimiter = '\n';
+    Storage().wordleGame = JsonUtils.encode(toJson());
 
   // Data Access
 
-  static Future<Set<String>?> loadDictionary() async {
-    dynamic content = await Content().loadContentItem('illordle_dictioary');
-    return SetUtils.from(JsonUtils.stringValue(content)?.split('\n'));
+  static Future<WordleDailyWord?> loadDailyWord() async {
+    WordleDailyWord? debugDailyWord = WordleDailyWord.fromJson(JsonUtils.decodeMap(Storage().debugWordleDailyWord));
+    return (debugDailyWord != null) ?  debugDailyWord : await loadDailyWordFromNet();
   }
 
-  static Future<WordleDailyWord?> loadDailyWord() async =>
-    Future.delayed(Duration(milliseconds: 500), () => _sampeDailtyWord);
+  static Future<WordleDailyWord?> loadDailyWordFromNet() async {
+    String? url = Config().illordleDailyWordUrl;
+    Response? response = (url?.isNotEmpty == true) ? await Network().get(url) : null;
+    return (response?.succeeded == true) ? WordleDailyWord.fromJson(JsonUtils.decodeMap(response?.body)) : null;
+  }
 
-  static WordleDailyWord _sampeDailtyWord = const WordleDailyWord(
-    word: 'viral',
-    date: '2025-11-14',
-    author: 'Anna Ceja',
-    quote: 'Viral TokTok star Joshua Block visits UI',
-  );
+  static Future<Set<String>?> loadDictionary() async {
+    String? url = Config().illordleWordsUrl;
+    Response? response = (url?.isNotEmpty == true) ? await Network().get(url) : null;
+    return (response?.succeeded == true) ? SetUtils.from(JsonUtils.stringValue(response?.body)?.split(RegExp(r'[\r\n]')).map((String word) => word.toUpperCase())) : null;
+  }
 }
 
 class WordleDailyWord {
   final String word;
-
-  final String? date;
+  final DateTime? dateUtc;
   final String? author;
-  final String? quote;
+  final String? storyTitle;
+  final String? storyUrl;
 
   const WordleDailyWord({
     required this.word,
-    this.date, this.author, this.quote,
+    this.dateUtc, this.author,
+    this.storyTitle, this.storyUrl,
   });
 
-  DateTime? get dateTime =>
-    (date != null) ? DateFormat('yyyy-MM-dd').tryParse(date ?? '') : null;
+  // Json Serialization
+  static WordleDailyWord? fromJson(Map<String, dynamic>? json) =>
+    _fromJsonWord(JsonUtils.stringValue(json?['word']), json: json);
+
+  static WordleDailyWord? _fromJsonWord(String? word, { Map<String, dynamic>? json }) => ((word != null) && word.isNotEmpty) ?
+    WordleDailyWord(
+      word: word.toUpperCase(),
+      dateUtc: dateFromString(JsonUtils.stringValue(json?['date'],)),
+      author: JsonUtils.stringValue(json?['author']),
+      storyTitle: JsonUtils.stringValue(json?['story_title']),
+      storyUrl: JsonUtils.stringValue(json?['story_url']),
+    ) : null;
+
+
+  Map<String, dynamic> toJson() => <String, dynamic>{
+    'word': word,
+    'date': dateAsString,
+    'author': author,
+    'story_title': storyTitle,
+    'story_url': storyUrl,
+  };
+
+  static DateTime? dateFromString(String? value) =>
+    (value != null) ? DateFormat(_dateFormat).tryParse(value, true) : null;
+
+  String? get dateAsString => (dateUtc != null) ?
+    (DateFormat(_dateFormat).format(dateUtc!) + ' GMT') : null;
+
+  static const String _dateFormat = 'EEE, d MMM yyyy hh:mm:ss Z';
+
+  // Accessories
+  int get wordLength => word.wordLength;
+  int get numberOfWords => word.numberOfWords;
+  double get asectRatio => word.asectRatio;
 
   // Equality
 
@@ -650,16 +862,23 @@ class WordleDailyWord {
   bool operator==(Object other) =>
     (other is WordleDailyWord) &&
     (word == other.word) &&
-    (date == other.date) &&
+    (dateUtc == other.dateUtc) &&
     (author == other.author) &&
-    (quote == other.quote);
+    (storyTitle == other.storyTitle);
 
   @override
   int get hashCode =>
     (word.hashCode) ^
-    (date?.hashCode ?? 0) ^
+    (dateUtc?.hashCode ?? 0) ^
     (author?.hashCode ?? 0) ^
-    (quote?.hashCode ?? 0);
+    (storyTitle?.hashCode ?? 0);
+
+}
+
+extension _WordleWordRules on String {
+  int get wordLength => length;
+  int get numberOfWords => wordLength + 1;
+  double get asectRatio => wordLength.toDouble() / numberOfWords.toDouble();
 }
 
 enum _WordleLetterStatus { inPlace, inUse, outOfUse }
@@ -678,4 +897,36 @@ enum _WordleLetterDisplay { move, rack, }
 
 extension _StringExt on String {
   bool get isAlpha => (length == 1) && (toUpperCase() != toLowerCase());
+}
+
+extension _LettersBag on Map<String, int> {
+
+  static Map<String, int> fromWord(String word) {
+    Map<String, int> bag = <String, int>{};
+    for (int index = 0; index < word.length; index++) {
+      String letter = word.substring(index, index + 1);
+      bag[letter] = (bag[letter] ?? 0) + 1;
+    }
+    return bag;
+  }
+
+  bool removeLetter(String letter) {
+    int letterCount = this[letter] ?? 0;
+    if (0 < letterCount) {
+      letterCount = letterCount - 1;
+      if (0 < letterCount) {
+        this[letter] = letterCount;
+      }
+      else {
+        this.remove(letter);
+      }
+      return true; // removed
+    }
+    else {
+      return false; // not in the bag
+    }
+  }
+
+  int _letterCount(String letter) => this[letter] ?? 0;
+  bool containsLetter(String letter) => (0 < _letterCount(letter));
 }